Add files via upload
Description of project idea and target audience:

Experiencing new environments always brings along thousands of unfamiliar challenges for every individual. These challenges become more striking when it comes to studying abroad, far from home. Yearly, thousands of students enter Germany from different countries with various backgrounds. Adapting to a new environment, would be the most significant factor that elevates personal success of these students, that is not possible but with gaining information.
To support our community of students at “TU Ilmenau”, we came up with the idea of having a social media application for new students who need information in different fields of living and studying in Germany. Residential matters, Language barrier, academic problems, cultural differences and dealing with bureaucracy in Germany are the main challenges that we identified through our target audience research. Therefore, our project is based on approaches to cope with these problems. Moreover, this application tries to act as a bridge between alumni and new students to spread the information and experiences of studying at TU Ilmenau.